+++ point.cpp
@@ -3,7 +3,7 @@
 #include "point.h"
 
 Point::Point():d_x(0), d_y(0){}
-Point::Point(int x, int y):d_x(x), d_y(y){}
+Point::Point(double x, double y):d_x(x), d_y(y){}
 
  void Point::afficherPoint(std::ostream& os){
     os << "( " << d_x << " , " << d_y << " )";
@@ -14,22 +14,22 @@ Point::Point(int x, int y):d_x(x), d_y(y){}
 
      is >> c >> d_x >> c >> d_y >> c;
  }
-   int Point::distance(const Point& p)
+   double Point::distance(const Point& p)
  {
-  int dx= p.d_x-d_x;
-  int dy=p.d_y-d_y;
+  double dx= p.d_x-d_x;
+  double dy=p.d_y-d_y;
   return std::sqrt(dx*dx +dy*dy);
  }
-  int Point::x(){
+  double Point::x(){
      return d_x;
  }
-  int Point::y(){
+  double Point::y(){
      return d_y;
  }
- void Point::setX(int x){
+ void Point::setX(double x){
      d_x = x;
  }
- void Point::setY(int y){
+ void Point::setY(double y){
      d_y = y;
  }
