+++ point.h
@@ -4,9 +4,6 @@
 #include<ostream>
 #include<istream>
 
-
-
-
 class Point
 {
     //pourquoi mettre les operateurs d'entree sortie en freind?
@@ -14,17 +11,17 @@ class Point
     friend std::istream& operator >>(std::istream &os, Point &p);
 public:
     Point();
-    Point(int x, int y);
+    Point(double x, double y);
      void afficherPoint(std::ostream&);
      void lirePoint(std::istream&);
-     int distance(const Point &p);
-     int x();
-     int y();
-     void setX(int x);
-     void setY(int y);
+     double distance(const Point &p);
+     double x();
+     double y();
+     void setX(double x);
+     void setY(double y);
 private:
-    int d_x;
-    int d_y;
+    double d_x;
+    double d_y;
 };
 /* Si > 0, alors C est à gauche de la droite, si < 0 alors C est à droite, sinon il est sur la droite.
 int aGauche(Point a, Point b, Point c);
